+++ app/src/renderer/components/common/TmConnectedNetwork.vue
@@ -30,7 +30,7 @@
     id="tm-disconnected-network"
     class="tm-connected-network tm-disconnected-network"
   >
-    <img class="tm-connected-network-loader" src="~assets/images/loader.svg">
+    <img class="tm-connected-network-loader" src="~assets/images/loader.svg" />
     <div
       v-tooltip.top="networkTooltip"
       class="
@@ -81,11 +81,11 @@ export default {
   justify-content: space-between;
   margin: 0.5rem;
   padding: 0.5rem;
-  position: absolute;
+  position: fixed;
   bottom: 0;
   left: 0;
   right: 0;
-  max-width: 240px;
+  max-width: 208px; /* sidebar width minus margin */
 }
 
 .tm-connected-network .chain-id {