+++ app/src/renderer/components/common/ToolBar.vue
@@ -9,11 +9,7 @@
       <i v-tooltip.bottom="'Refresh'" class="material-icons">refresh</i>
     </a>
     <slot />
-    <router-link
-      v-if="session.signedIn"
-      id="settings"
-      to="/preferences"
-    >
+    <router-link v-if="session.signedIn" id="settings" to="/preferences">
       <i v-tooltip.bottom="'Preferences'" class="material-icons">settings</i>
     </router-link>
     <a
@@ -21,13 +17,16 @@
       id="signOut-btn"
       @click="signOut()"
     >
-      <i v-tooltip.bottom.end="'Sign Out'" class="material-icons">
+      <i
+        v-tooltip.bottom.end="'Sign Out'"
+        class="material-icons"
+      >
         exit_to_app
       </i>
     </a>
     <tm-btn
       v-if="!session.signedIn"
-      id="signIn-btn"
+      class="sign-in-button"
       value="Sign In"
       color="primary"
       @click.native="signIn()"
@@ -66,8 +65,10 @@ export default {
 .tool-bar {
   display: flex;
   align-items: center;
-  height: fit-content;
-  padding: 1rem 0 0;
+}
+
+.sign-in-button {
+  padding-left: 1.5rem;
 }
 
 .tool-bar a {