+++ app/src/renderer/components/governance/PageProposal.vue
@@ -1,10 +1,5 @@
 <template>
   <tm-page data-title="Proposal">
-    <template slot="menu-body">
-      <tm-balance />
-      <tool-bar />
-    </template>
-
     <tm-data-error v-if="!proposal" />
 
     <template v-else>
@@ -153,8 +148,6 @@ import BigNumber from "bignumber.js"
 import { mapGetters } from "vuex"
 import num from "scripts/num"
 import TmBtn from "common/TmBtn"
-import ToolBar from "common/ToolBar"
-import TmBalance from "common/TmBalance"
 import TmDataError from "common/TmDataError"
 import TextBlock from "common/TextBlock"
 import ModalDeposit from "./ModalDeposit"
@@ -163,11 +156,9 @@ import TmPage from "common/TmPage"
 export default {
   name: `page-proposal`,
   components: {
-    TmBalance,
     TmBtn,
     ModalDeposit,
     ModalVote,
-    ToolBar,
     TmDataError,
     TmPage,
     TextBlock
@@ -211,16 +202,18 @@ export default {
         .toNumber()
     },
     yesPercentage({ tally, totalVotes } = this) {
-      return num.percentInt(tally.yes / totalVotes)
+      return num.percentInt(totalVotes === 0 ? 0 : tally.yes / totalVotes)
     },
     noPercentage({ tally, totalVotes } = this) {
-      return num.percentInt(tally.no / totalVotes)
+      return num.percentInt(totalVotes === 0 ? 0 : tally.no / totalVotes)
     },
     noWithVetoPercentage({ tally, totalVotes } = this) {
-      return num.percentInt(tally.no_with_veto / totalVotes)
+      return num.percentInt(
+        totalVotes === 0 ? 0 : tally.no_with_veto / totalVotes
+      )
     },
     abstainPercentage({ tally, totalVotes } = this) {
-      return num.percentInt(tally.abstain / totalVotes)
+      return num.percentInt(totalVotes === 0 ? 0 : tally.abstain / totalVotes)
     },
     tally({ proposals, proposalId } = this) {
       const { yes, no, abstain, no_with_veto } =