+++ app/src/renderer/components/transactions/LiBankTransaction.vue
@@ -7,26 +7,21 @@
     <template v-if="sent">
       <div slot="caption">
         Sent&nbsp;
-        <b>{{ atoms(coins.amount) }}</b>
+        <b>{{ full(atoms(coins.amount)) }}</b>
         <span>&nbsp;{{ coins.denom.toUpperCase() }}</span>
       </div>
-      <span
-        slot="details"
-      >
-        <template
-          v-if="sentSelf"
-        >
+      <span slot="details">
+        <template v-if="sentSelf">
           To yourself!
-        </template><template
-          v-else
-        >
+        </template><template v-else>
           To {{ receiver }}
         </template>
       </span>
     </template><template v-else>
       <div slot="caption">
+
         Received&nbsp;
-        <b>{{ atoms(coins.amount) }}</b>
+        <b>{{ full(atoms(coins.amount)) }}</b>
         <span>&nbsp;{{ coins.denom.toUpperCase() }}</span>
       </div>
       <span slot="details">From {{ sender }}</span>
@@ -36,7 +31,7 @@
 
 <script>
 import LiTransaction from "./LiTransaction"
-import { atoms } from "../../scripts/num.js"
+import { atoms, full } from "../../scripts/num.js"
 
 export default {
   name: `li-bank-transaction`,
@@ -52,7 +47,8 @@ export default {
     }
   },
   data: () => ({
-    atoms
+    atoms,
+    full
   }),
   computed: {
     tx() {