+++ app/src/renderer/components/transactions/LiGovTransaction.vue
@@ -7,7 +7,7 @@
     <template v-if="txType === `cosmos-sdk/MsgSubmitProposal`">
       <div slot="caption">
         Submit {{ tx.proposal_type.toLowerCase() }} proposal&nbsp;
-        <b>{{ atoms(tx.initial_deposit[0].amount) }}</b>
+        <b>{{ full(atoms(tx.initial_deposit[0].amount)) }}</b>
         <span>&nbsp;{{ tx.initial_deposit[0].denom }}s</span>
       </div>
       <div slot="details">
@@ -18,7 +18,7 @@
       <div slot="caption">
         Deposit&nbsp;
         <template>
-          <b>{{ pretty(atoms(tx.amount[0].amount)) }}</b>
+          <b>{{ full(atoms(tx.amount[0].amount)) }}</b>
           <span>&nbsp;{{ tx.amount[0].denom }}s</span>
         </template>
       </div>
@@ -45,7 +45,7 @@
 
 <script>
 import LiTransaction from "./LiTransaction"
-import { pretty, atoms } from "../../scripts/num.js"
+import { full, atoms } from "../../scripts/num.js"
 
 export default {
   name: `li-gov-transaction`,
@@ -69,7 +69,7 @@ export default {
     }
   },
   data: () => ({
-    pretty,
+    full,
     atoms
   }),
   computed: {